\docType{data}
\name{YU61female}
\alias{YU61female}
\title{Two-region observed population characteristics for Yugoslavian females, 1961.}
\format{A data frame with 36 rows and 7 variables}
\description{
A dataset containing observed 1961 female population,
births, deaths and migration between two regions, Slovenia
and the rest of Yugoslavia. Original study is, analyzed as
demonstration dataset in Rogers (1975) and Willekens &
Rogers (1978).
}
\keyword{datasets}

